+++ docs/locales/th/guide/changelog.md
@@ -1,5 +1,20 @@
 # บันทึกการเปลี่ยนแปลง
 
+## 2.23.0 (2025-11-120)
+
+- คุณสมบัติ:
+  - สร้างคอมโพเนนต์ Radio Grouped พร้อมการผูก v-model และการสนับสนุนอาร์เรย์ตัวเลือก
+  - เพิ่มการรองรับ Display Helper แบบกำหนดเองสำหรับคอมโพเนนต์ Input Contact Number
+  - ลบการทดสอบ e2e
+    ([#35957](https://dev.azure.com/sproutphil/Sprout%20Design%20System/_git/Sprout%20Design%20System%20Next/pullrequest/36031) โดย @jarias)
+
+## 2.22.3 (2025-11-19)
+
+- แก้ไข:
+  - เพิ่มการรองรับ Display Helper สำหรับ Radio Component
+  - อัปเดตเอกสารประกอบสำหรับ Radio Component
+    ([#35957](https://dev.azure.com/sproutphil/Sprout%20Design%20System/_git/Sprout%20Design%20System%20Next/pullrequest/36031) โดย @jarias)
+
 ## 2.22.2 (2025-11-19)
 
 - แก้ไข: